Solve problem 007 (#8)
* Method to check whether a number is a prime number

* Method to get prime number by index

This algorithm is not ideal. Actually, it is quite insufficient since it
does extra operations. There must be a more suitable algorithm, but I
don't know know how to implement it. Thus, I made this one which does
the job in a brute-force mode.

Also, add two tests for sample input and for task input.

* Remove unnecessary comment

* Use Java 11