Solve Problem 12 with prime factorization
If the number is big, then the most efficient way of finding the number
of divisors of that number is by getting the prime factorization of the
number. Then, get the exponents of those prime factorization, add +1 to
each of exponents and multiple numbers together. This way, the task is
solver within less than 1 sec.