Solve Problem 015 - number of unique paths
Apply an efficient algorithm that calculates the number of unique paths
in a square grid.
First step is to fill in first row and first column with one (1).
Next, start row by row. Value in a next cell will be the sum of the cell
 above and a cell to the left. And so on.
The resulting number in a bottom right cell is a number of unique paths
from top left to bottom right.

This method will work for a square 2d array. The task asks for a
solution of the 20x20 grid. It won't work as is for a 2d array that is
not square. For loop range should be adjusted. I didn't make it because
it was asked in the task.