Solve Problem 22
1. Split the string with names into a list, strip all double quotes.
2. Calculate score of a name by converting a string into a stream of
chars. Each char has a numeric representation, that is a sequential
order from A to Z. Thus, subtract the 'A' character from a given
character and add 1 to get the position relative to 'A'.
3. Get an index of the name from the list. Indices in the list are
zero-based, thus need to add 1 to the `indexOf()` return value. Multiply
the index of the name with its score.
4. Sum all numbers.